+++ lib/views/subscription_details_page.dart
@@ -20,6 +20,7 @@ class SubscriptionDetailsPageWidgetState extends State<SubscriptionDetailsPage> 
   final theme = new Themes();
   List charities = [];
   String _selectedCharity;
+  int _selectedIndex = -1;
 
   @override
   void initState() {
@@ -29,14 +30,23 @@ class SubscriptionDetailsPageWidgetState extends State<SubscriptionDetailsPage> 
 
   void _setupCharities() async {
     charities = await data.charities();
-    _selectedCharity = subscription.charity;
+    setState(() {
+      _selectedCharity = subscription.charity;
+    });
   }
 
   Widget build(BuildContext context) {
     return new Scaffold(
       appBar: _buildBar(context),
       body: new Container(
-        child: new Text('Hi', style: theme.textStyle,),
+        child: new Column(
+          children: <Widget>[
+            _buildFirstRow(),
+            _buildSecondRow(),
+            _buildDropDown(),
+            _buildDeleteButton(),
+          ],
+        ),
       ),
     );
   }
@@ -52,7 +62,152 @@ class SubscriptionDetailsPageWidgetState extends State<SubscriptionDetailsPage> 
         style: theme.textStyle,
       ),
       centerTitle: true,
+      actions: <Widget>[
+        new IconButton(icon: theme.editIcon, onPressed: _editPressed)
+      ],
     );
+  }
+
+  Widget _buildFirstRow() {
+    return new Container(
+      child: new Row(
+        children: <Widget>[
+          new Expanded(
+            child: new Text(
+              'Player: ${subscription.name}',
+              style: theme.textStyle,
+            ),
+          ),
+          new Expanded(
+            child: new Text(
+              'Charity: ${subscription.charity}',
+              style: theme.textStyle,
+            ),
+          )
+        ],
+      ),
+      padding: EdgeInsets.only(
+        top: 50.0,
+        left: 40.0,
+      ),
+    );
+  }
+
+  Widget _buildSecondRow() {
+    String subscribedSince = subscription.time.substring(0, 15);
+    subscribedSince.trim();
+    return new Container(
+      child: new Row(
+        children: <Widget>[
+          new Expanded(
+            child: new Text(
+              'Goals: ${subscription.goals}',
+              style: theme.textStyle,
+            ),
+          ),
+          new Expanded(
+            child: new Text(
+              'Subscribed Since: $subscribedSince.',
+              style: theme.textStyle,
+            ),
+          )
+        ],
+      ),
+      padding: EdgeInsets.only(
+        top: 50.0,
+        left: 40.0,
+      ),
+    );
+  }
+
+  Widget _buildDropDown() {
+    return new Container(
+      child: new DropdownButton(
+          items: new List<DropdownMenuItem>.generate(charities.length ?? 0, (int index) {
+            return new DropdownMenuItem(
+              child: new Text(charities[index]['name']),
+              value: index,
+            );
+          }),
+          hint: new Text(
+            _selectedCharity ?? 'Select a Charity',
+            style: theme.textStyle,
+          ),
+          onChanged: (value) {
+            setState(() {
+              _selectedIndex = value;
+              _selectedCharity = charities[value]['name'];
+            });
+          },
+      ),
+      padding: EdgeInsets.only(
+        top: 100.0,
+      ),
+    );
+  }
+
+  Widget _buildDeleteButton() {
+    return new Container(
+      child: RaisedButton(
+        color: Colors.red,
+        textColor: Colors.white,
+        child: Text('Delete Subscription'),
+        onPressed: _deleteSubscription,
+      ),
+      padding: EdgeInsets.only(
+        top: 200.0,
+      ),
+    );
+  }
+
+  void _deleteSubscription()  {
+    print('ok');
+    showDialog(
+        context: context,
+        builder: _showAlert
+    );
+  }
+
+  Widget _showAlert(BuildContext context) {
+    return new AlertDialog(
+      content: new Text(
+        'Are you sure you want to delete this subscription?',
+      ),
+      actions: <Widget>[
+        new FlatButton(
+          child: Text('Delete', style: TextStyle(color: Colors.red),),
+          onPressed: _confirmDelete,
+        ),
+        new FlatButton(
+          child: Text('Cancel'),
+          onPressed: () => Navigator.pop(context),
+        )
+      ],
+    );
+  }
+
+  void _editPressed() async {
+
+    String name = charities[_selectedIndex]['name'];
+    String id = charities[_selectedIndex]['id'];
+    if (_selectedCharity != subscription.charity) {
+      bool result = await sub.updateSubscription(subscription.id, name, id);
+      if (result) {
+        setState(() {
+          _selectedIndex = -1;
+          subscription.charity = name;
+          subscription.charityId = id;
+        });
+      }
+    }
+  }
+
+  void _confirmDelete() async {
+    bool result = await sub.removeSubscription(subscription.id);
+    if (result) {
+      Navigator.of(context).pop();
+      Navigator.of(context).pop(true);
+    }
   }
 
 }